Reesctructuracion del apartado de establecimiento, separando la logica de subida de imagenes, se estructrua el apartado de contratos, incluyendo dos nuevos campos

+++ Back/WebGESCOMPAH/Entity/DTOs/Implements/Business/EstablishmentDto/EstablishmentSelectDto.cs
@@ -7,7 +7,7 @@ namespace Entity.DTOs.Implements.Business.EstablishmentDto
     {
         public string Name { get; set; } = null!;
         public string Description { get; set; } = null!;
-        public double AreaM2 { get; set; }
+        public decimal AreaM2 { get; set; }
         public decimal RentValueBase { get; set; }
         public string Address { get; set; } = string.Empty;
         public int PlazaId { get; set; }
